requirements.txt
mcp>=1.0.0
cohere>=5.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
